+++ apps/pub_social/bin/init.dart
@@ -1,0 +1,3 @@
+void main() {
+  print("Dart is running");
+}