+++ packages/jaspr/lib/src/server/run_app.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:shelf/shelf.dart';
@@ -10,7 +11,6 @@ import 'server_app.dart';
 import 'server_handler.dart';
 
 /// Main entry point on the server
-/// TODO: Add hint about usage of global variables and isolate state
 void runApp(Component app) {
   _checkInitialized('runApp');
   ServerApp.run(_createSetup(app));
@@ -29,18 +29,24 @@ Handler serveApp(AppHandler handler) {
   });
 }
 
-/// A record containing the status code, body, and headers for a response.
-typedef ResponseLike = ({int statusCode, ResponseBody body, Map<String, List<String>> headers});
-
 /// Directly renders the provided component to HTML. Returns a [ResponseLike] object.
 ///
 /// - Accepts a [request] object for getting the current url and headers.
-/// - When [standalone] is false (default), the html output will have a full document structure (html, head, body).
-Future<ResponseLike> renderComponent(Component app, {Request? request, bool standalone = false}) async {
-  _checkInitialized('renderComponent');
+/// - When [standalone] is false (default), the html output will always have a full document structure (html, head, body).
+Future<ResponseLike> renderApp(Component app, {Request? request, bool standalone = false}) async {
+  _checkInitialized('renderApp');
   request ??= Request('GET', Uri.parse('https://0.0.0.0/'));
   var fileLoader = proxyFileLoader(request, staticFileHandler());
   return render(_createSetup(app), request, fileLoader, standalone);
+}
+
+/// Directly renders the provided component to HTML. Returns a [String].
+///
+/// - Accepts a [request] object for getting the current url and headers.
+/// - When [standalone] is false (default), the html output will have a full document structure (html, head, body).
+Future<String> renderComponent(Component app, {Request? request, bool standalone = false}) async {
+  final response = await renderApp(app, request: request, standalone: standalone);
+  return response.body.content;
 }
 
 void _checkInitialized(String method) {
@@ -58,15 +64,21 @@ SetupFunction _createSetup(Component app) {
   };
 }
 
+/// A record containing the status code, body, and headers for a response.
+typedef ResponseLike = ({int statusCode, ResponseBody body, Map<String, List<String>> headers});
 
-
+/// A class representing the body of a response. It can be either a string or bytes.
 sealed class ResponseBody {
   const factory ResponseBody(String content) = ResponseBodyString;
+  const factory ResponseBody.bytes(List<int> bytes) = ResponseBodyBytes;
+
+  String get content;
 }
 
 class ResponseBodyString implements ResponseBody {
   const ResponseBodyString(this.content);
 
+  @override
   final String content;
 }
 
@@ -74,4 +86,7 @@ class ResponseBodyBytes implements ResponseBody {
   const ResponseBodyBytes(this.bytes);
 
   final List<int> bytes;
+
+  @override
+  String get content => utf8.decode(bytes);
 }