+++ packages/jaspr_og/example/jaspr_og_example.dart
@@ -1,5 +1,10 @@
+import 'dart:io';
+
+import 'package:jaspr/jaspr.dart';
 import 'package:jaspr_og/jaspr_og.dart';
 
-void main() {
-  
+void main() async {
+  final image = await renderSvg(svg([]), width: 1200, height: 630);
+
+  File('image.png').writeAsBytesSync(image);
 }